feat: implemented persistance
* chore: setting up persistence layer, connection with local database

* chore: continuing setup and connection with database

* chore: finishing table setup

* Created Review daos and fixed some Ids

* Merge branch 'master' into persistance

# Conflicts:
#	models/src/main/java/ar/edu/itba/paw/dtos/OrderDirection.java
#	models/src/main/java/ar/edu/itba/paw/models/User.java
#	persistence/src/main/java/ar/edu/itba/paw/persistence/ReviewDaoImpl.java

* Fixed merge

+++ models/src/main/java/ar/edu/itba/paw/models/Review.java
@@ -5,7 +5,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
 public class Review {
-    private final Integer id;
+    private final Long id;
     private final User author;
     private final String title;
     private final String content;
@@ -13,7 +13,7 @@ public class Review {
     private final Integer rating;
     private final Game reviewedGame;
 
-    public Review(Integer id, User author, String title, String content, LocalDate created, Integer rating, Game reviewedGame) {
+    public Review(Long id, User author, String title, String content, LocalDate created, Integer rating, Game reviewedGame) {
         this.id = id;
         this.author = author;
         this.title = title;
@@ -23,7 +23,7 @@ public class Review {
         this.reviewedGame = reviewedGame;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
